requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
av==12.3.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.8
ctranslate2==4.5.0
distro==1.9.0
enum34==1.1.10
faster-whisper==1.0.3
filelock==3.16.1
flatbuffers==24.3.25
fsspec==2024.10.0
h11==0.14.0
halo==0.0.31
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
Jinja2==3.1.4
jiter==0.7.1
joblib==1.4.2
log-symbols==0.0.14
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
onnxruntime==1.20.0
openai==1.54.4
openwakeword==0.6.0
packaging==24.2
protobuf==5.28.3
psutil==6.1.0
pvporcupine==1.9.5
PyAudio==0.2.14
pycaw==20240210
pydantic==2.9.2
pydantic_core==2.23.4
pypiwin32==223
pyreadline3==3.5.4
python-vlc==3.0.21203
pyttsx3==2.98
pywin32==308
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.5.0
six==1.16.0
sniffio==1.3.1
SpeechRecognition==3.11.0
spinners==0.0.24
sympy==1.13.1
termcolor==2.5.0
threadpoolctl==3.5.0
tokenizers==0.20.3
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
webrtcvad-wheels==2.0.14
websocket-client==1.8.0
websockets==12.0